Add test for simple add. Change number to int because floats might give long answers

+++ s_expr_int_calculator/s_expr_int_calculator/token.hpp
@@ -7,7 +7,7 @@
 namespace sx {
 using FilePosition = uint16_t;
 using TokenLength = uint8_t;
-using TokenNumber = double;
+using TokenNumber = int64_t;
 
 enum class TokenType : uint8_t { IDENTIFIER, NUMBER, ERROR };
